fix zy

+++ sdks/zy/src/org/yunyue/GameProxyImpl.java
@@ -55,7 +55,7 @@ public class GameProxyImpl extends GameProxy{
 
     public void pay(Activity activity, String ID, String name, String orderID, float price, String callBackInfo, JSONObject roleInfo, final PayCallBack payCallBack) {
         PayParams params=new PayParams();
-        params.setAmount(price);
+        params.setAmount((int)price);
         params.setPropsName("元宝");
         params.setOrderId(orderID);
         params.setExtraParam(callBackInfo);
